Added lines with "speed of dark" and "zamboni".
Added a few lines. Mentioned "the plural form of zamboni" without going into too much detail and spoiling the magic. Indeed, poetry con oft be a collaborative process between the writer and the reader. The reader can ask themselves the kind of questions that can one day grow them into a poet themselves. Anything you're passionate about is an excuse to strive for excellence! Sometimes commit summaries can be considered literature. And if it's not, I hope that a long description like this can at least make someone laugh with the absurdity of it all.