Poet: Kamala Das
This is an autobiographical poem which throws light on the life and work of Kamala Das. Das begins the poem by saying that she doesn’t understand politics but she knows the name of politicians probably referring to the fact that power is in the hands of a few elites and that it is usually males who run the country. She then gives a brief introduction of herself before she focuses on English being the medium she uses to expresses herself; how people criticize her for that; and why it is no one’s business other than herself. The poem then moves to her early and unsuccessful marriage and how the society she lives in is male dominated. The primary focus of the poem is the situation of women in a patriarchal society and the unjust burdens women have to go through in this male dominated world. Kamala Das or Kamala Surayya is one of the best known Indian female poets and this is her most famous work.